# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TRACE workflow adapter",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sessions/current": {
      "delete": {
        "summary": "Clear Session",
        "operationId": "clear_session_api_v1_sessions_current_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_v1_uploads_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_uploads_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs": {
      "post": {
        "summary": "Run",
        "operationId": "run_api_v1_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          },
          {
            "name": "x-openai-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Openai-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TextSummary"
                  },
                  {
                    "$ref": "#/components/schemas/YoutubeSummary"
                  },
                  {
                    "$ref": "#/components/schemas/ArticleSummary"
                  },
                  {
                    "$ref": "#/components/schemas/FileSummary"
                  },
                  {
                    "$ref": "#/components/schemas/OCR"
                  },
                  {
                    "$ref": "#/components/schemas/ImageGeneration"
                  },
                  {
                    "$ref": "#/components/schemas/Speech"
                  },
                  {
                    "$ref": "#/components/schemas/Captions"
                  },
                  {
                    "$ref": "#/components/schemas/Code"
                  },
                  {
                    "$ref": "#/components/schemas/Content"
                  },
                  {
                    "$ref": "#/components/schemas/Research"
                  },
                  {
                    "$ref": "#/components/schemas/Intelligence"
                  },
                  {
                    "$ref": "#/components/schemas/DocumentQA"
                  },
                  {
                    "$ref": "#/components/schemas/URLQA"
                  },
                  {
                    "$ref": "#/components/schemas/Support"
                  }
                ],
                "discriminator": {
                  "propertyName": "tool",
                  "mapping": {
                    "text-summary": "#/components/schemas/TextSummary",
                    "youtube-summary": "#/components/schemas/YoutubeSummary",
                    "article-summary": "#/components/schemas/ArticleSummary",
                    "file-summary": "#/components/schemas/FileSummary",
                    "ocr": "#/components/schemas/OCR",
                    "image": "#/components/schemas/ImageGeneration",
                    "speech": "#/components/schemas/Speech",
                    "captions": "#/components/schemas/Captions",
                    "code": "#/components/schemas/Code",
                    "content": "#/components/schemas/Content",
                    "research": "#/components/schemas/Research",
                    "documents": "#/components/schemas/Intelligence",
                    "document-qa": "#/components/schemas/DocumentQA",
                    "url-qa": "#/components/schemas/URLQA",
                    "support": "#/components/schemas/Support"
                  }
                },
                "title": "Inputs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request-bound events: started, stage, warning, result, error. These are workflow observations, not model tokens.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexts/documents": {
      "post": {
        "summary": "Documents",
        "operationId": "documents_api_v1_contexts_documents_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          },
          {
            "name": "x-openai-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Openai-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentsContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request-bound events: started, stage, warning, result, error. These are workflow observations, not model tokens.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexts/urls": {
      "post": {
        "summary": "Urls",
        "operationId": "urls_api_v1_contexts_urls_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          },
          {
            "name": "x-openai-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Openai-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/URLsContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request-bound events: started, stage, warning, result, error. These are workflow observations, not model tokens.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexts/{context_id}/query": {
      "post": {
        "summary": "Query",
        "operationId": "query_api_v1_contexts__context_id__query_post",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          },
          {
            "name": "x-openai-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Openai-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Query"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request-bound events: started, stage, warning, result, error. These are workflow observations, not model tokens.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "summary": "Artifact",
        "operationId": "artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArticleSummary": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "article-summary",
            "title": "Tool"
          },
          "url": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "url"
        ],
        "title": "ArticleSummary"
      },
...
      "Body_upload_api_v1_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_uploads_post"
      },
      "Captions": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "captions",
            "title": "Tool"
          },
          "url": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "url"
        ],
        "title": "Captions"
      },
...
      "Code": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "code",
            "title": "Tool"
          },
          "prompt": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Prompt"
          },
          "language": {
            "type": "string",
            "enum": [
              "python",
              "javascript",
              "java",
              "c",
              "c++"
            ],
            "title": "Language",
            "default": "python"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "prompt"
        ],
        "title": "Code"
      },
...
      "Content": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "content",
            "title": "Tool"
          },
          "idea": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Idea"
          },
          "tone": {
            "type": "string",
            "enum": [
              "professional",
              "playful",
              "educational",
              "launch-ready",
              "social-first"
            ],
            "title": "Tone",
            "default": "professional"
          },
          "platforms": {
            "items": {
              "type": "string",
              "enum": [
                "linkedin",
                "x",
                "instagram",
                "youtube",
                "blog"
              ]
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Platforms",
            "default": [
              "linkedin",
              "x"
            ]
          },
          "include_audio": {
            "type": "boolean",
            "title": "Include Audio",
            "default": false
          },
          "voice": {
            "type": "string",
            "enum": [
              "alloy",
              "echo",
              "fable",
              "onyx",
              "nova",
              "shimmer"
            ],
            "title": "Voice",
            "default": "alloy"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "idea"
        ],
        "title": "Content"
      },
...
      "DocumentQA": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "document-qa",
            "title": "Tool"
          },
          "context_id": {
            "type": "string",
            "title": "Context Id"
          },
          "question": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "context_id",
          "question"
        ],
        "title": "DocumentQA"
      },
...
      "DocumentsContext": {
        "properties": {
          "file_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "File Ids"
          }
        },
        "type": "object",
        "required": [
          "file_ids"
        ],
        "title": "DocumentsContext"
      },
      "FileSummary": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "file-summary",
            "title": "Tool"
          },
          "file_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "File Ids"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "file_ids"
        ],
        "title": "FileSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageGeneration": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "image",
            "title": "Tool"
          },
          "prompt": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "prompt"
        ],
        "title": "ImageGeneration"
      },
      "Intelligence": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "documents",
            "title": "Tool"
          },
          "file_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "File Ids"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "file_ids"
        ],
        "title": "Intelligence"
      },
      "OCR": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "ocr",
            "title": "Tool"
          },
          "file_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "File Ids"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "file_ids"
        ],
        "title": "OCR"
      },
...
      "Query": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "Query"
      },
      "Research": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "research",
            "title": "Tool"
          },
          "topic": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Topic"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "topic",
          "urls"
        ],
        "title": "Research"
      },
...
      "Speech": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "speech",
            "title": "Tool"
          },
          "text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Text"
          },
          "voice": {
            "type": "string",
            "enum": [
              "alloy",
              "echo",
              "fable",
              "onyx",
              "nova",
              "shimmer"
            ],
            "title": "Voice",
            "default": "alloy"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "text"
        ],
        "title": "Speech"
      },
      "Support": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "support",
            "title": "Tool"
          },
          "question": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Question"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "question",
          "urls"
        ],
        "title": "Support"
      },
...
      "TextSummary": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "text-summary",
            "title": "Tool"
          },
          "text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "text"
        ],
        "title": "TextSummary"
      },
      "URLQA": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "url-qa",
            "title": "Tool"
          },
          "context_id": {
            "type": "string",
            "title": "Context Id"
          },
          "question": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "context_id",
          "question"
        ],
        "title": "URLQA"
      },
      "URLsContext": {
        "properties": {
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "urls"
        ],
        "title": "URLsContext"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YoutubeSummary": {
        "properties": {
          "tool": {
            "type": "string",
            "const": "youtube-summary",
            "title": "Tool"
          },
          "url": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "url"
        ],
        "title": "YoutubeSummary"
      }
    }
  }
}